refactor tests after deprecations remove
