+++ tests/Unit/Aggregation/Bucketing/GlobalAggregationTest.php
@@ -9,9 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace ONGR\ElasticsearchDSL\Tests\Unit\Aggregation;
+namespace ONGR\ElasticsearchDSL\Tests\Unit\Bucketing\Aggregation;
 
-use ONGR\ElasticsearchDSL\Aggregation\GlobalAggregation;
+use ONGR\ElasticsearchDSL\Aggregation\Bucketing\GlobalAggregation;
 
 class GlobalAggregationTest extends \PHPUnit_Framework_TestCase
 {
