+++ tests/Unit/Aggregation/Metric/PercentilesAggregationTest.php
@@ -9,9 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace ONGR\ElasticsearchDSL\Tests\Unit\Aggregation;
+namespace ONGR\ElasticsearchDSL\Tests\Unit\Metric\Aggregation;
 
-use ONGR\ElasticsearchDSL\Aggregation\PercentilesAggregation;
+use ONGR\ElasticsearchDSL\Aggregation\Metric\PercentilesAggregation;
 
 class PercentilesAggregationTest extends \PHPUnit_Framework_TestCase
 {
